+++ app/src/main/java/ru/netology/nmedia/dto/Post.kt
@@ -3,18 +3,18 @@ package ru.netology.nmedia.dto
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class Post (
+data class Post(
     var id: Long,
     val author: String,
     val content: String,
     val published: String,
     var likedByMe: Boolean = false,
     var likes: Int = 0,
-    var countLike: Double,
-    var countShare: Double,
-    var countEye: Double,
-    val videoUrl: String?,
+    var countLike: Int,
+    var countShare: Int,
+    var countEye: Int,
+    val videoUrl: String?
 
-) {
+    ) {
         val videoAttachment = videoUrl != null
 }
